Feat: deleteComment 권한 확인 (#205)

+++ src/main/java/travelplanner/project/demo/domain/comment/service/CommentService.java
@@ -1,5 +1,6 @@
 package travelplanner.project.demo.domain.comment.service;
 
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
@@ -11,10 +12,12 @@ import travelplanner.project.demo.domain.comment.dto.request.CommentCreateReques
 import travelplanner.project.demo.domain.comment.dto.request.CommentEditRequest;
 import travelplanner.project.demo.domain.comment.dto.response.CommentResponse;
 import travelplanner.project.demo.domain.comment.repository.CommentRepository;
+import travelplanner.project.demo.domain.member.domain.Member;
 import travelplanner.project.demo.domain.post.post.domain.Post;
 import travelplanner.project.demo.domain.post.post.repository.PostRepository;
 import travelplanner.project.demo.global.exception.ApiException;
 import travelplanner.project.demo.global.exception.ErrorType;
+import travelplanner.project.demo.global.util.AuthUtil;
 import travelplanner.project.demo.global.util.PageUtil;
 
 import java.util.ArrayList;
@@ -25,6 +28,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CommentService {
 
+    private final AuthUtil authUtil;
     private final PostRepository postRepository;
     private final CommentRepository commentRepository;
 
@@ -85,12 +89,18 @@ public class CommentService {
     }
 
     @Transactional
-    public void deleteComment(Long postId, Long commentId) {
+    public void deleteComment(Long postId, Long commentId, HttpServletRequest request) {
         postRepository.findById(postId)
                 .orElseThrow(() -> new ApiException(ErrorType.POST_NOT_FOUND));
 
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new ApiException(ErrorType.COMMENT_NOT_FOUND));
+
+        Member currentMember = authUtil.getCurrentMember(request);
+
+        if (comment.getPost().getMember() != currentMember) {
+            throw new ApiException(ErrorType.USER_NOT_AUTHORIZED);
+        }
 
         commentRepository.delete(comment);
     }
